ART: Selectivity Cleanup patch
As per comments on patchsets 7 and 15  on
https://android.intel.com/#/c/325723/ this commit addresses those
comments.

Changed dex limit to be 64 bit.
Fixed identation.
Added consistent linkage for rest of selectivity members.

Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-30584
Category: device enablement
Domain: AOSP.ART-ME
Origin: internal
Upstream-Candidate: no, proprietary
Depends-Change-Id: I1ced77b8194688611149739fad16d7ac18bed46c
Reviewed-on: https://android.intel.com:443/350748

+++ compiler/dex/selectivity.h
@@ -151,8 +151,8 @@ namespace Selectivity {
    * @brief Getter and Setter for the dex opcode limit for the application.
    */
   #define SELECTIVITY_DEFAULT_PER_APP_DEX_LIMIT 600000
-  void SetPerAppDexLimit(uint32_t dex_limit);
-  uint32_t GetPerAppDexLimit();
+  void SetPerAppDexLimit(uint64_t dex_limit);
+  uint64_t GetPerAppDexLimit();
 }  // namespace Selectivity
 }  // namespace art
 #endif  // ART_COMPILER_DEX_SELECTIVITY_H_
